+++ NEWS.md
@@ -7,6 +7,10 @@
 * Histogram metrics now have a built-in `time()` method for collecting
   observations on the duration of an expression.
 
+* The built-in `process_cpu_seconds_total` metric is now correctly reported as
+  a Counter instead of a Gauge. This fixes an inconsistency with other
+  Prometheus clients that would prevent metrics being used with Pushgateway.
+
 # openmetrics 0.1.1
 
 * Fixes inaccurate CPU usage when using `register_default_metrics()`.